Rename alert-sending command
